The more a man is imbued with the ordered regularity of all events, the firmer becomes his conviction that there is no room left by the side of this ordered regularity for causes of a different nature.  For him neither the rule of human nor the rule of divine will exists as an independent cause of natural events.  To be sure, the doctrine of a personal God interfering with natural events could never be refuted, in the real sense, by science, for this doctrine can always take refuge in those domains in which scientific knowledge has not yet been able to set foot.
But I am persuaded that such behavior on the part of the representatives
of religion would not only be unworthy but also fatal.  For a doctrine which
is able to maintain itself not in clear light, but only in the dark, will
of necessity lose its effect on mankind, with incalculable harm to human
progress.  In their struggle for the ethical good, teachers of religion
must have the stature to give up the doctrine of a personal God, that is,
give up that source of fear and hope which in the past placed such vast
powers in the hands of priests.  In their labors they will have to avail
themselves of those forces which are capable of cultivating the Good, the
True, and the Beautiful in humanity itself.  This is, to be sure, a more
difficult but an incomparably more worthy task.
- Albert Einstein